message overview...nice

+++ calvi_table/Message.swift
@@ -13,7 +13,7 @@ struct Message {
     var name: String
     var message: String
     var url: String
-    var date: Int32
+    var date: Double
     var idFrom: String
     var idTo: String
     var articleId: Int64
